Add hello world widget support
