+++ 4 - WS Part 3/src/HelloWorldGraphical.java
@@ -12,6 +12,12 @@ public class HelloWorldGraphical{
         windowSystem = new WindowSystem(1024, 768);
 
         //add windows
-        windowSystem.addNewWindow(0.2f, 0.1f, 400, 500, "Hello world and don't panic");
+        SimpleWindow newWindow = windowSystem.addNewWindow(0.2f, 0.1f, 400, 500, "Hello world and don't panic");
+        RatHelloWorld widget = new RatHelloWorld();
+        widget.test = "lalala";
+        newWindow.setWidget(widget);
+        RatHelloWorld testWidget = (RatHelloWorld)newWindow.getWidget();
+        System.out.println("banana");
+        System.out.println(widget.test);
     }
 }
