+++ 4 - WS Part 3/src/RatHelloWorld.java
@@ -1,5 +1,6 @@
 public class RatHelloWorld extends RATWidget {
 
+	public String test;
 	public RatHelloWorld(){
     	//list initialization
     	super();
